[FIX] account: discrepancy between amount_residual and amount_residual_signed
Fixes an issue where amount_residual and amount_residual_signed did not have
the same absolute value when multiple currencies are involved on an account move.

This issue was seen when viewing the "Amount Due" (amount_residual) on a customer
invoice form view and comparing it to the "Amount Due" either on the list view
or on the Aged Payables financial report. In cases where multiple currencies were
involved on the customer invoice and customer payment, the amount_residual_signed
field showed an incorrect non-zero amount for fully paid invoices.